feat: move getAuthIdByAuthMethod to lit-auth-client
